Update README.md
This project demonstrates a basic email spam classifier using the Naive Bayes algorithm. The model is trained on a small dataset of email texts labeled as "Spam" or "Not Spam". It utilizes CountVectorizer to convert the text data into a numerical format, which is then used to train the model. The classifier is evaluated on a test set, and its performance is measured in terms of accuracy.

Features:
Data Preprocessing: Text data is converted into numerical data using CountVectorizer.
Model Training: The classifier is trained using the MultinomialNB algorithm.
Prediction: The model predicts whether an email is spam or not based on its content.
How to Use:
Train the model using a labeled email dataset.
Test the model with unseen data (new emails).
Predict whether new emails are spam or not.